Enabled by default, no default creds, move class, fix typo in test
