+++ spring-boot-admin-docs/src/main/asciidoc/security.adoc
@@ -63,7 +63,7 @@ You can specify credentials via configuration properties in your admin server.
 
 TIP: You can use this in conjuction with https://cloud.spring.io/spring-cloud-kubernetes/1.1.x/reference/html/#secrets-propertysource[spring-cloud-kubernetes^] to pull credentials from https://kubernetes.io/docs/concepts/configuration/secret/[secrets^].
 
-To enable pulling credentials from properties you must set `spring.boot.admin.instance-auth.enabled` to `true`.
+To enable pulling credentials from properties the `spring.boot.admin.instance-auth.enabled` property must be `true` (default).
 
 NOTE: If your clients provide credentials via metadata (i.e., via service annotations), that metadata will be used instead of the properites.
 
@@ -77,7 +77,10 @@ spring.boot.admin:
     default-user-name: "${some.user.name.from.secret}"
     default-user-password: "${some.user.password.from.secret}"
     service-map:
-      my-sb-admin-service:
+      my-first-service-to-monitor:
+        user-name: "${some.user.name.from.secret}"
+        user-password: "${some.user.password.from.secret}"
+      my-second-service-to-monitor:
         user-name: "${some.user.name.from.secret}"
         user-password: "${some.user.password.from.secret}"
 ----
